energy: Adding electricity reporting using northq
This lua script does report the electrical consumption splitted in two
graphs.

France EDF provider is providing low_cost period & expensive periods of time everyday.
This script is reporting the power consumption for both by detecting the
period of time reported by a block scene.

Note this script is using the thingspeak web service for the graphing to
make the code lighter & easier to bootstrap.