dsaudio: Don't fail if no song in the playlist
While trying to get the current status of the player, if no song is
present the current code was generating the following exception :

	[ERROR] 20:07:33: line 774: attempt to index field 'song' (a userdata value)

This patch ensuure that the song name exist before using it. If no song
is present, then we just return an empty string.

Thanks for users reporting that case I never add on mine.