dsaudio: Don't fail if no song in playlist
When the playlist was empty, the data type of song is no more an array
but a userdata type which makes the code properly failing.

That patch is so assuming that
- if the song is a userdata, then it means there is no song in the playlist.
- if the song is not a userdata, then it's an array featuring songs

That's pretty crap and could change with further releases of DSAudio
(from Synology). We might have to add more hacks here if required.