dsaudio: Consider retrying if response lenght > 53000
By multiple testing, it have been found that if the response lenght is
greater then 53000 chars, the lua function manipulating that string is
immediately returning.

I don't really know if this is related to a lua or a fibaro restriction.

This workaround consider the response isn't correct if the string lenght
is too long. This will make the retry code being active and a new
request is issued.

This will make the get_random_song() code a little bit more robust.